join functional in component Timer
