+++ src/components/Timer/CountdownToEndYear.jsx
@@ -8,15 +8,23 @@ const CountdownToEndYear = ({ startDateTime, endDateTime }) => {
   );
   if (days + hours + minutes + seconds <= 0) {
     console.log("countdown is over");
-    return <ShowCounter days="00" hours="00" minutes="00" seconds="00" />;
+    return (
+      <>
+        <p>До закінчення року залишилось</p>
+        <ShowCounter days="00" hours="00" minutes="00" seconds="00" />
+      </>
+    );
   } else {
     return (
-      <ShowCounter
-        days={days}
-        hours={hours}
-        minutes={minutes}
-        seconds={seconds}
-      />
+      <>
+        <p>До закінчення року залишилось</p>
+        <ShowCounter
+          days={days}
+          hours={hours}
+          minutes={minutes}
+          seconds={seconds}
+        />
+      </>
     );
   }
 };
