+++ src/components/Timer/getEndOfYear.js
@@ -1,0 +1,5 @@
+const getEndOfYear = (year) => {
+  return new Date(year, 11, 32);
+};
+
+export default getEndOfYear;
