+++ src/components/Timer/getPadTime.js
