+++ src/components/Timer/transformMSTime.js
